task: обновления функионала

+++ pkg/logger/message.go
@@ -3,6 +3,6 @@ package logger
 type Message struct {
 	Message     string
 	FullMessage string
-	Datetime    int
+	Datetime    int64
 	ExtraData   map[string]any
 }
